requirements.txt
google-cloud==0.34.0
google-api-python-client==1.7.9
google-cloud-pubsub==0.42.1
pyyaml==5.1.1